+++ app.js
@@ -3,19 +3,17 @@ const server = express()
 
 server.use(express.urlencoded({ extended: true }));
 
-const jerseysController = require("./controllers/jerseys")
-
+const jerseysController = require("./controllers/jerseysController")
+const generalPagesController = require("./controllers/generalPagesController")
 
 server.use(express.static("public"))
 
-server.get("/", jerseysController.getHomePage)
+server.get("/", generalPagesController.getHomePage)
+server.post('/submit', generalPagesController.handleFormSubmission)
+
 server.get("/getAllJerseys", jerseysController.getAllJerseys)
 server.get("/getJersey", jerseysController.getJersey)
 server.get("/deleteJersey", jerseysController.deleteJersey)
-
-
-server.post('/submit', jerseysController.handleFormSubmission)
-
 
 
 server.listen(81)
